feat: repl - mobile fixes
