+++ hooks/index.ts
@@ -8,3 +8,4 @@ export { default as useSaveAndProceed } from './useSaveAndProceed'
 export { default as useSaveAndReturn } from './useSaveAndReturn'
 export { default as useLessonStatus } from './useLessonStatus'
 export { default as useLessonUnlocked } from './useLessonUnlocked'
+export { default as useDynamicHeight } from './useDynamicHeight'
